clustermole: Unbiased Single-Cell Transcriptomic Data Cell Type Identification

Assignment of cell type labels to single-cell RNA sequencing (scRNA-seq) clusters is often a time-consuming process that involves manual inspection of the cluster marker genes complemented with a detailed literature search. This is especially challenging when unexpected or poorly described populations are present. The clustermole R package provides methods to query thousands of human and mouse cell identity markers sourced from a variety of databases.

See also

Author

Maintainer: Igor Dolgalev igor.dolgalev@nyumc.org (ORCID)

Authors:

Igor Dolgalev igor.dolgalev@nyumc.org (ORCID)