Cell types based on the expression of all genes

Perform enrichment of cell type signatures based on the full gene expression matrix.

clustermole_enrichment(expr_mat, species, method = "gsva")

Arguments

expr_mat
Expression matrix (logCPMs, logFPKMs, or logTPMs) with genes as rows and clusters/populations/samples as columns.

species
Species: hs for human or mm for mouse.

method
Enrichment method: ssgsea, gsva, singscore, or all. The method to use for the estimation of gene set enrichment scores. The options are ssGSEA (Barbie et al, 2009), GSVA (Hänzelmann et al, 2013), singscore (Foroutan et al, 2018), or a combination of all three methods.

Value

A data frame of enrichment results.

References

Barbie, D., Tamayo, P., Boehm, J. et al. Systematic RNA interference reveals that oncogenic KRAS-driven cancers require TBK1. Nature 462, 108–112 (2009). https://doi.org/10.1038/nature08460

Hänzelmann, S., Castelo, R. & Guinney, J. GSVA: Gene set variation analysis for microarray and RNA-Seq data. BMC Bioinformatics 14, 7 (2013). https://doi.org/10.1186/1471-2105-14-7

Foroutan, M., Bhuva, D.D., Lyu, R. et al. Single sample scoring of molecular phenotypes. BMC Bioinformatics 19, 404 (2018). https://doi.org/10.1186/s12859-018-2435-4

Examples

# my_enrichment <- clustermole_enrichment(expr_mat = my_expr_mat, species = "hs")